% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  ...,
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/"
)
}
\arguments{
\item{...}{Additional options passed to `golem_opts`}

\item{onStart}{Startup function}

\item{options}{Options list passed to shinyApp}

\item{enableBookmarking}{Bookmarking mode}

\item{uiPattern}{URL pattern for the UI}
}
\description{
Run the Shiny Application
}
